requirements.txt
# AWS SDK
boto3>=1.34.0

# Slack SDK
slack-sdk>=3.27.0

# GitHub API
PyGithub>=2.3.0

# Google Calendar API
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.119.0

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Date/Time handling
python-dateutil>=2.8.2
pytz>=2024.1

# AI/LLM (OpenAI)
openai>=1.12.0

# Testing
pytest>=8.0.0
moto>=5.0.0
